craete checkout page and order controller

+++ database/migrations/2024_09_17_151444_create_carts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_id');  // Reference to the product
             $table->unsignedBigInteger('user_id')->nullable();  // Nullable for unauthenticated users
             $table->uuid('cookie_id')->nullable();  // Nullable for unauthenticated users
-            $table->integer('quantity');
+            $table->unsignedSmallInteger('quantity');
             $table->string('color');
             $table->string('size');
             $table->timestamps();
